Split out firefly interface
