+++ private-data-transfer-cli/src/index.ts
@@ -1,81 +1,6 @@
-import axios, { AxiosInstance } from 'axios';
-import WebSocket from 'ws';
+import { FireFly, FireFlyListener, FireFlyData } from "./firefly";
 
 const TIMEOUT = 15 * 1000;
-
-interface FireFlyData {
-  value: string;
-}
-
-interface FireFlyDataIdentifier {
-  id: string;
-  hash: string;
-}
-
-interface FireFlyMessage {
-  id: string;
-  type: string;
-  message: {
-    data: FireFlyDataIdentifier[];
-  }
-}
-
-class FireFlyListener {
-  private ws: WebSocket;
-  private connected: Promise<void>;
-  private messages: FireFlyMessage[] = [];
-
-  constructor(port: number, namespace='default') {
-    this.ws = new WebSocket(`ws://localhost:${port}/ws?namespace=${namespace}&ephemeral&autoack`);
-    this.connected = new Promise<void>(resolve => {
-      this.ws.on('open', resolve);
-      this.ws.on('message', (data: string) => {
-        this.messages.push(JSON.parse(data));
-      });
-    });
-  }
-
-  ready() {
-    return this.connected;
-  }
-
-  close() {
-    this.ws.close();
-  }
-
-  firstMessageOfType(type: string, timeout: number) {
-    return new Promise<FireFlyMessage | undefined>(async (resolve) => {
-      const expire = Date.now() + timeout;
-      while (Date.now() < expire) {
-        for (const message of this.messages) {
-          if (message.type === type) {
-            resolve(message);
-            return;
-          }
-        }
-        await new Promise(resolve => setTimeout(resolve, 100));
-      }
-      return undefined;
-    });
-  }
-}
-
-class FireFly {
-  private rest: AxiosInstance;
-
-  constructor(port: number) {
-    this.rest = axios.create({ baseURL: `http://localhost:${port}/api/v1` });
-  }
-
-  async sendBroadcast(data: FireFlyData[]) {
-    await this.rest.post('/namespaces/default/broadcast/message', { data });
-  }
-
-  async getData(id: string) {
-    const response = await this.rest.get<FireFlyData>(`/namespaces/default/data/${id}`)
-    return response.data;
-  }
-}
 
 async function main() {
   const firefly1 = new FireFly(5000);
